Added Title bars name #139 Fixed user has to refresh after importing keys #135
