+++ apps/provider/src/app/admin/(internal)/groups/page.tsx
@@ -1,36 +1,25 @@
-'use client';
-
+import type { Metadata } from 'next'
 import React from 'react'
-import {Button} from "@igniter/ui/components/button";
-import {AddOrUpdateAddressGroupDialog} from "@/components/AddOrUpdateAddressGroupDialog";
 import AddressGroupsTable from "./table";
-import {useQueryClient} from "@tanstack/react-query";
+import { GetAppName } from '@/actions/ApplicationSettings'
+import AddNewAddressGroup from '@/app/admin/(internal)/groups/AddNew'
+
+export async function generateMetadata(): Promise<Metadata> {
+  const appName = await GetAppName()
+
+  return {
+    title: `Addresses Group - ${appName}`,
+    description: "Light up your earnings with Igniter",
+  }
+}
 
 export default function GroupsPage() {
-  const queryClient = useQueryClient();
-
-  const [isAddingGroup, setIsAddingGroup] = React.useState(false);
   return (
     <div className="flex flex-col gap-10">
-      {isAddingGroup && (
-        <AddOrUpdateAddressGroupDialog
-          onClose={(shouldRefreshAddressGroups) => {
-            setIsAddingGroup(false);
-            if (shouldRefreshAddressGroups) {
-              queryClient.invalidateQueries({queryKey: ['groups']});
-            }
-          }}
-        />
-      )}
       <div className="mx-30 py-10">
         <div className={'flex flex-row items-center gap-4'}>
           <h1>Address Groups</h1>
-          <Button
-            variant={"outline"}
-            onClick={() => setIsAddingGroup(true) }
-          >
-            Add New
-          </Button>
+          <AddNewAddressGroup />
         </div>
         <div className="container mx-auto ">
           <AddressGroupsTable />
