+++ apps/provider/src/app/admin/(internal)/miners/page.tsx
@@ -1,36 +1,26 @@
-'use client';
-
+import type { Metadata } from 'next'
 import React from 'react'
-import {Button} from "@igniter/ui/components/button";
-import {AddOrUpdateRelayMinerDialog} from "@/components/AddOrUpdateRelayMinerDialog";
 import RelayMinersTable from "./table";
-import {useQueryClient} from "@tanstack/react-query";
+import { GetAppName } from '@/actions/ApplicationSettings'
+import AddNewRelayMiner from '@/app/admin/(internal)/miners/AddNew'
+
+export async function generateMetadata(): Promise<Metadata> {
+    const appName = await GetAppName()
+
+    return {
+        title: `Relay Miners - ${appName}`,
+        description: "Light up your earnings with Igniter",
+    }
+}
 
 export default function RelayMinersPage() {
-    const queryClient = useQueryClient();
-
-    const [isAddingMiner, setIsAddingMiner] = React.useState(false);
     return (
         <div className="flex flex-col gap-10">
-            {isAddingMiner && (
-                <AddOrUpdateRelayMinerDialog
-                    onClose={(shouldRefreshRelayMiners) => {
-                        setIsAddingMiner(false);
-                        if (shouldRefreshRelayMiners) {
-                            queryClient.invalidateQueries({queryKey: ['relay-miners']});
-                        }
-                    }}
-                />
-            )}
+
             <div className="mx-30 py-10">
                 <div className={'flex flex-row items-center gap-4'}>
                     <h1>Relay Miners</h1>
-                    <Button
-                        variant={"outline"}
-                        onClick={() => setIsAddingMiner(true) }
-                    >
-                        Add New
-                    </Button>
+                    <AddNewRelayMiner />
                 </div>
                 <div className="container mx-auto ">
                     <RelayMinersTable />
